Reset
I reset the git repository to clear the OCR XML files from the version
history. These aren't needed for the app to function, and are only
needed to create the index.

The page OCR is available at the Internet Archive and can be downloaded
with this command, which talks to the Airtable database and downloads
what has been digitized from the Internet Archive:

    npm run fetch-docs

Similarly the episode WEBVTT files are stored in an S3 bucket and are
downloaded when you run this command:

    npm run fetch-episodes