Back to Search
This commit adds functionality to change the URL when a search is
performed, and then to introspect on it when visiting the page to
execute a search. This allows the back button to work in search
results: so when you click on a search result for details you can go
back to your results.

Also included is a small change to include facet-label-* classes for
facets: https://app.asana.com/0/1119509223481634/1147523936362465/f