Materials and Programs pages
This commit adds the /materials/ and /programs/ pages. I moved the
listing of series that was previously in /series/ into /programs/.
Unfortunately, series is used in the underlying Airtable data model, and
as a result is in quite a bit of the backend code. I haven't updated all
the places that talk about series in the code, but at the surface level
in the HTML and the URLs 'program' is used in place of 'series'.

At some point (when there's time) we should refactor everything on the backend
to match how we talk about things in the interface. Just to keep
ourselves sane.