Series searching
This commit adds search by series title to the search page. It also
includes a change to the flexsearch index that minimizes the amount of
data that is stored, in order to save on index size. This means that the OCR
text needs to be saved on the filesystem and fetched as needed when
displaying search results. Additionally the search now will load more
results if they are available when the user scrolls to the bottom of the
page (infinite scroll).