Contents List
This commit adds some new functionality to create a series / box / folder
finding aid when collecting documents, which gets serialized as
static/data/finding-aid.json.

The Explore Related Materials page uses this JSON structure to populate
the Contents List. At the moment it only includes materials that we have
digitized, rather than being a complete list of the contents of the
collection.