Folder Descriptions
This commit adds folder descriptions to the Explore Materials page. I
needed to query and store the descriptions in
static/data/finding-aid.json, which required re-running the fetch-docs
command. To speed this process up, which normally tries to download ocr
for every folder/item I added --skip-ocr to skip the download of ocr
when running fetch-docs.