Trailing periods in subjects
This commit removes trailing periods from subjects when the
fetch-authorities commmad is run. It seems like we have inconsistent
data entry on subject names in Airtable. Orginarily this wouldn't be a
problem but subject names form part of links in canned searches and some
content management systems (e.g. Asana) will treat a trailng period as
extraneous. Since an exact match is required for facet searches the
missing period resulted in an empty search result page.

Also included here is a favicon for the site since I noticed it was
defaulting to a Wordpress icon.