Gatsby v2.22.20
This commit includes an upgrade to Gatsby v2.22.20 which seemed to fix
our dissappearing series pages. There were some minor template changes
that needed to be made, as well as updated Mirador importing. In
addition this commit includes the result of running npm run fetch-data
to get the latest data from Airtable and the Internet Archive, as well
as a reindex.