Fullscreen document
This commit removes the two column layout for documents and puts allows
the Mirador viewer to use all the width of the space, and moves the
metadata down below.

In addition I set the Mirador fullscreen configuration option to true to allow users to view the documents using the entire screen.