graphql ignore snac data
I noticed that the build was taking a really long time after the
static/data/snac data was added. I think that's because the
gatsby-source-filesystem and gatsby-transformer-json were trying to
determine the schema of these files and make them available for graphql.
Since we don't really need graphql on these just yet, and may never need
it, this commit asks gatsby-source-filesystem to ignore the snac data
like it does the ocr in static/data/ocr.