gatsby-plugin-offline
This commit adds gatsby-plugin-offline which uses a service
worker to cache pages in the browser. The plugin is basically a wrapper
for adding Workbox [1] to a Gatsby site. At the moment only top level menu
pages (About, Exhibits, Search, etc) are pre-cached. However Workbox
will also cache other pages on demand, including files like the large
index.json that is needed for searching. This means that the index.json
file will be fetched from local storage rather than the network when
possible.

Previously the index was being pre-loaded in gatsby-browser.js which meant
that any page load would trigger the fetching the large JSON file.
Sometimes this would cause the also page rendering to freeze while the JSON was fetched and parsed. Now the index.json is loaded by the SearchPage component in
src/pages/search.js. It was converted from a functional component to a class
component so that the fetch of index.json could happen as part of
componentDidMount.

I've tested updating the index.json file to make sure that changes are
handled properly. It appears that Workbox will serve stale data until
new data has been fetched, which seems pretty nice. In addition there is
a little hook to ask users to reload the application if the service
worker changes.

[1] https://developers.google.com/web/tools/workbox/