Render TimelineJS
This commit adds the timelinejs npm module and imports it into the
VisualizationsPage component. I found some notes on how to wire things up
in this issue:

https://github.com/NUKnightLab/TimelineJS3/issues/633

Since I'm not familiar with the original it's hard for me to say if it's working
completely, but at least it appears to be loading.

Since TimelineJS CSS uses less I needed to install & add
gatsby-plugin-less. Also, in order to get the build to work I needed to
remove timelinejs from the server side build since it wanted to access
the global navigator which is only available in the browser:

https://www.gatsbyjs.com/docs/debugging-html-builds/#fixing-third-party-modules