Fix for the cache + graphql error
This fixes the problem we started seeing recently when running `npm run
start` more than once caused a Missing field GraphQL error, which
resulted in the site failing to run and the GraphQL editor to show that
the wikipedia.json file was no longer queryable. Running `npm run clean`
would then allow things to work again. But this caused problems in
environments when we wanted to keep the cache around for a faster build,
and it was just annoying when developing the site.

I tried a bunch of things but ran across an issue on GitHub where
someone said that changing the name of a property from `identifier` to
`name` in the JSON data they were querying made this problem go away. I
didn't have an `identifier` property but I did have an `id` and thought
changing it to `personId` might help. It did make the problem go away. I
tried to document this on GitHub for Gatsby here:

https://github.com/gatsbyjs/gatsby/issues/11747#issuecomment-877641772

This was a difficult problem to debug, and the solution was really
unsatisfying because I don't understand why it helped. I think if we are
going to continue using Gatsby in the future we might want to invest
some time learning how to define our schema rather than letting Gatsby
infer it.