showThemes
This commit adds a showThemes method on the Search page, which is pushed
down into Search component to get called with true or false depending on
whether there are any results or not. This allows the themes to show up
only when there are no results.

In addition this commit also updates the browser location when facets
are selected, so searches are shareable as URLs.