Fixing unstable exhibit lede display
It appears that the lede sometimes appears and sometimes does not when
visiting the production site, and any static build. Under the devserver
everything seems fine.

I think that the cause for this is using dangerouslySetInnerHTML to set
HTML that has nested <p> tags?!?!

https://github.com/gatsbyjs/gatsby/issues/2750#issuecomment-342245034

I've replaced the <p> tag with a <div> for the lede and will see if this
fixes things at www.unlockingtheairwaves.org like it did for my local
build.