Normalize JSON files to match Airtable tables
This commit addes the Associated Place(s) and Associated Subject(s) fields to the CPF display, the latter which had values linked to a canned search.

fetch-authorities.js was updated to write the JSON files names so they match the Airtable names. This means that people.json is now cpf.json, and wikidata.json is cpf-pages.json. The relevant GraphQL queries have been updated to target the new data, and templates adjusted to use it. Finally the /about/ page was updated to list the new files instead of the old ones.